+++ prospr/core/src/protein.hpp
@@ -48,10 +48,10 @@ class Protein {
         int get_score();
 
         /* Returns the number of performed changes. */
-        int get_solutions_checked();
+        unsigned long long get_solutions_checked();
 
         /* Returns the number of amino acids placed. */
-        int get_aminos_placed();
+        unsigned long long get_aminos_placed();
 
         /* Returns if the amino acid at the given index is weighted. */
         bool is_weighted(int index);
@@ -97,8 +97,8 @@ class Protein {
         int last_move;
         std::vector<int> last_pos;
         int score;
-        int aminos_placed;
-        int solutions_checked;
+        unsigned long long aminos_placed;
+        unsigned long long solutions_checked;
         std::vector<AminoAcid*> amino_acids;
 
         /* Change score according to the already performed addition or removal of the
